Refactor Shared Project to PCL XamarinStore.Core
I had already did some of the work of converting some of the shared
project to be more PCL friendly , specially the webclient stack. But i
found out that there was a PCL branch already on the project, so i took
some of the changes that wer there by James Clancey, special the
PCLStorae, MD5, gravatar and small changes in the Android and IOS
client.

No need to link files now, all the data access is done in the Core
project and shared by Android/IOS/Windows